+++ packages/core/src/js/behaviors.js
@@ -36,6 +36,7 @@ import { installSplitter } from './splitter.js';
 import { installShell } from './shell.js';
 import { installDatagrid } from './datagrid.js';
 import { installValidation } from './validation.js';
+import { installThemeToggle } from './theme-toggle.js';
 import { installFieldErrors } from './field-errors.js';
 
 function init() {
@@ -68,6 +69,7 @@ function init() {
   installShell();
   installDatagrid();
   installValidation();
+  installThemeToggle();
   installFieldErrors();
 }
 
@@ -109,6 +111,7 @@ export {
   installShell,
   installDatagrid,
   installValidation,
+  installThemeToggle,
   installFieldErrors,
 };